Mavenize module, and fix a few compile errors to enable module to build with maven successfully.

+++ api/src/main/java/org/openmrs/module/mohbilling/db/hibernate/HibernateBillingDAO.java
@@ -13,23 +13,9 @@
  */
 package org.openmrs.module.mohbilling.db.hibernate;
 
-import java.math.BigDecimal;
-import java.text.DateFormat;
-
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashSet;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.hibernate.Criteria;
-
-import org.hibernate.Query;
 import org.hibernate.SQLQuery;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -37,7 +23,6 @@ import org.hibernate.criterion.Expression;
 import org.hibernate.criterion.Order;
 import org.hibernate.criterion.Restrictions;
 import org.openmrs.Concept;
-
 import org.openmrs.Patient;
 import org.openmrs.User;
 import org.openmrs.api.context.Context;
@@ -59,7 +44,17 @@ import org.openmrs.module.mohbilling.model.Recovery;
 import org.openmrs.module.mohbilling.model.ServiceCategory;
 import org.openmrs.module.mohbilling.model.ThirdParty;
 import org.openmrs.module.mohbilling.service.BillingService;
-import org.springframework.orm.hibernate3.SessionFactoryUtils;
+
+import java.math.BigDecimal;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 
 /**
@@ -92,8 +87,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getInsurance(org.openmrs.
-	 *      module.mohbilling.model.Insurance)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getInsurance(Integer)
 	 */
 	@Override
 	public Insurance getInsurance(Integer insuranceId) {
@@ -110,8 +104,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getInsurancePolicy(org.openmrs
-	 *      .module.mohbilling.model.InsurancePolicy)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getInsurancePolicy(Integer)
 	 */
 	@Override
 	public InsurancePolicy getInsurancePolicy(Integer cardId) {
@@ -146,8 +139,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getBillableServiceByConcept(Concept
-	 *      concept, Insurance insurance)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getBillableServiceByConcept(FacilityServicePrice price, Insurance insurance)
 	 */
 	@Override
 	public BillableService getBillableServiceByConcept(
@@ -171,8 +163,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getPatientBill(org.openmrs
-	 *      .module.mohbilling.model.PatientBill)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getPatientBill(Integer)
 	 */
 	@Override
 	public PatientBill getPatientBill(Integer billId) {
@@ -184,8 +175,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getPatientBill(org.openmrs
-	 *      .module.mohbilling.model.PatientBill)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getPatientBill(Integer)
 	 */
 	@Override
 	public ThirdParty getThirdParty(Integer thirdPartyId) {
@@ -197,8 +187,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#saveThirdParty(org.openmrs
-	 *      .module.mohbilling.model.ThirdParty)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#saveThirdParty(ThirdParty)
 	 */
 	@Override
 	public void saveThirdParty(ThirdParty thirdParty) {
@@ -209,8 +198,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#saveInsurance(org.openmrs
-	 *      .module.mohbilling.model.Insurance)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#saveInsurance(Insurance)
 	 */
 	@Override
 	public void saveInsurance(Insurance insurance) {
@@ -221,8 +209,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#saveInsurancePolicy(org.openmrs
-	 *      .module.mohbilling.model.InsurancePolicy)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#saveInsurancePolicy(InsurancePolicy)
 	 */
 	@Override
 	public void saveInsurancePolicy(InsurancePolicy card) {
@@ -234,8 +221,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#savePatientBill(org.openmrs
-	 *      .module.mohbilling.model.PatientBill)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#savePatientBill(PatientBill)
 	 */
 	@Override
 	public void savePatientBill(PatientBill bill) {
@@ -412,8 +398,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#buildCohort(org.openmrs.module.mohbilling.model.Insurance,
-	 *      java.util.Date, java.util.Date, java.lang.Integer, java.lang.String)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#billCohortBuilder(Insurance, Date, Date, Integer, String, String, String)
 	 */
 	@Override
 	public List<PatientBill> billCohortBuilder(Insurance insurance,
@@ -583,7 +568,7 @@ public class HibernateBillingDAO implements BillingDAO {
 			psBill.setPatientServiceBillId((Integer) obj[0]);
 			psBill.setServiceDate((Date) obj[1]);
 			psBill.setUnitPrice((BigDecimal) obj[2]);
-			psBill.setQuantity((Integer) obj[3]);
+			psBill.setQuantity((BigDecimal) obj[3]);
 			psBill.setServiceOther((String) obj[4]);
 			psBill.setServiceOtherDescription((String) obj[5]);
 			psBill.setCreatedDate((Date) obj[6]);
@@ -814,8 +799,7 @@ public class HibernateBillingDAO implements BillingDAO {
 	/**
 	 * (non-Javadoc)
 	 * 
-	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getBillableServiceByConcept(Concept
-	 *      concept, Insurance insurance)
+	 * @see org.openmrs.module.mohbilling.db.BillingDAO#getServiceCategoryByName(String, Insurance)
 	 */
 	@Override
 	public ServiceCategory getServiceCategoryByName(String name,
